+++ wikitextparser/_spans.py
@@ -123,7 +123,8 @@ EXTENSION_TAGS_FINDITER = regex_compile(  # noqa
 # HTML tags
 # Tags:
 # https://infra.spec.whatwg.org/#ascii-whitespace
-SPACE_CHARS = rb' \t\n\u000C\r'  # \s - \v
+# \0 was added as a special case for wikitextparser
+SPACE_CHARS = rb' \t\n\u000C\r\0'  # \s - \v
 # http://stackoverflow.com/a/93029/2705757
 # chrs = (chr(i) for i in range(sys.maxunicode))
 # control_chars = ''.join(c for c in chrs if unicodedata.category(c) == 'Cc')